+++ packaging/windows/setup_cross_compilation_env.sh
@@ -51,10 +51,10 @@ function update_mingw_cross_env {
   cat > settings.mk <<EOF
 MXE_TARGETS = ${HOST}
 MXE_USE_CCACHE =
-MXE_PLUGIN_DIRS += plugins/gcc9
+MXE_PLUGIN_DIRS += plugins/gcc10
 JOBS = ${PARALLEL}
 
-MKVTOOLNIX_DEPENDENCIES=gettext libiconv zlib boost file flac ogg pthreads vorbis cmark libdvdread
+MKVTOOLNIX_DEPENDENCIES=gettext libiconv zlib boost flac ogg pthreads vorbis cmark libdvdread gmp
 MKVTOOLNIX_DEPENDENCIES+=qtbase qttranslations qtwinextras
 
 LOCAL_PKG_LIST=\$(MKVTOOLNIX_DEPENDENCIES)