+++ src/common/avc_es_parser.cpp
@@ -421,7 +421,7 @@ es_parser_c::handle_sps_nalu(memory_cptr const &nalu) {
       && sps_info.ar_found
       && (0 != sps_info.par_den)) {
     m_par_found = true;
-    m_par       = int64_rational_c(sps_info.par_num, sps_info.par_den);
+    m_par       = mtx_mp_rational_t(sps_info.par_num, sps_info.par_den);
   }
 }
 
@@ -940,7 +940,7 @@ es_parser_c::has_par_been_found()
   return m_par_found;
 }
 
-int64_rational_c const &
+mtx_mp_rational_t const &
 es_parser_c::get_par()
   const {
   assert(m_avcc_ready && m_par_found);
@@ -958,8 +958,8 @@ es_parser_c::get_display_dimensions(int width,
   if (0 >= height)
     height = get_height();
 
-  return std::make_pair<int64_t, int64_t>(1 <= m_par ? std::llround(width * boost::rational_cast<double>(m_par)) : width,
-                                          1 <= m_par ? height                                                    : std::llround(height / boost::rational_cast<double>(m_par)));
+  return std::make_pair<int64_t, int64_t>(1 <= m_par ? std::llround(width * static_cast<double>(m_par)) : width,
+                                          1 <= m_par ? height                                           : std::llround(height / static_cast<double>(m_par)));
 }
 
 size_t